requirements-test.txt
pytest==7.2.1
pytest-mock==3.10.0
python-dotenv==1.0.0
